+++ FitCloudKit/FitCloudKit.xcframework/ios-arm64_x86_64-simulator/FitCloudKit.framework/Headers/FitCloudFirmwareVersionObject.h
@@ -428,4 +428,7 @@
 /// 是否支持 App 助眠分析
 @property(nonatomic, readonly) BOOL shouldShowSleepAnalysisInApp;
 
+/// 是否支持外部存储
+@property(nonatomic, readonly) BOOL withExternalStorage;
+
 @end